requirements.txt
d2to1
pbr
future
fabric
docopts
flask
requests
sphinx
sphinx_bootstrap_theme
google-api-python-client
apiclient
httplib2
sphinxcontrib-exceltable
sphinxcontrib-httpdomain
pytimeparse
timestring
cookiecutter
urllib3
tzlocal
pytz
python-dateutil
